requirements.txt
fastapi
uvicorn
slowapi
uvicorn
mysqlclient
passlib
